dictionary

+++ src/main/java/at/abcdef/memmaster/controllers/dto/FolderDTO.java
@@ -39,5 +39,5 @@ public class FolderDTO {
     @Getter(lazy=true)
     private final List<FolderDTO> children = new ArrayList<>();
 
-    private Integer dictionary_count;
+    private Long dictionary_count;
 }
